+++ packages/terra-worklist-data-grid/CHANGELOG.md
@@ -11,6 +11,7 @@
 
 * Fixed
   * Reduced number of unnecessary rerenders in the `terra-worklist-data-grid` component.
+  * Fixed cells not honoring the value of props passed by consumers.
 
 ## 0.4.0 - (August 10, 2023)